+++ Runtime/Stores/Android/GooglePlay/AAR/Interfaces/IBillingClient.cs
@@ -3,14 +3,10 @@ using System.Collections.Generic;
 using UnityEngine.Purchasing.Models;
 using UnityEngine.Purchasing.Utils;
 
-namespace UnityEngine.Purchasing.Interfaces
+namespace UnityEngine.Purchasing.GoogleBilling.Interfaces
 {
-    interface IGoogleBillingClient
+    interface IBillingClient : IBillingClientBase
     {
-        void StartConnection(IBillingClientStateListener billingClientStateListener);
-        void EndConnection();
-        bool IsReady();
-        GoogleBillingConnectionState GetConnectionState();
         void QueryPurchasesAsync(string skuType, Action<IGoogleBillingResult, IEnumerable<AndroidJavaObject>> onQueryPurchasesResponse);
         void QueryProductDetailsAsync(List<string> skus, string type, Action<IGoogleBillingResult, List<AndroidJavaObject>> onProductDetailsResponseAction);
         AndroidJavaObject LaunchBillingFlow(AndroidJavaObject productDetails, string oldPurchaseToken, GooglePlayReplacementMode? replacementMode);