+++ Runtime/Stores/Android/GooglePlay/AAR/Models/LinkType.cs
@@ -1,0 +1,13 @@
+namespace UnityEngine.Purchasing.GoogleBilling.Models
+{
+    /// <summary>
+    /// Values from Java class LaunchExternalLinkParams.LinkType
+    /// <a href="https://developer.android.com/reference/com/android/billingclient/api/LaunchExternalLinkParams.LinkType"/>
+    /// </summary>
+    public enum LinkType
+    {
+        LINK_TYPE_UNSPECIFIED = 0,
+        LINK_TO_DIGITAL_CONTENT_OFFER = 1,
+        LINK_TO_APP_DOWNLOAD = 2
+    }
+}